+++ src/main/groovy/com/bmuschko/gradle/docker/tasks/image/DockerBuildImage.groovy
@@ -16,15 +16,11 @@
 package com.bmuschko.gradle.docker.tasks.image
 
 import com.bmuschko.gradle.docker.DockerRegistryCredentials
-import com.bmuschko.gradle.docker.response.ResponseHandler
-import com.bmuschko.gradle.docker.response.image.BuildImageResponseHandler
 import com.bmuschko.gradle.docker.tasks.AbstractDockerRemoteApiTask
 import com.bmuschko.gradle.docker.tasks.RegistryCredentialsAware
 import org.gradle.api.tasks.*
 
 class DockerBuildImage extends AbstractDockerRemoteApiTask implements RegistryCredentialsAware {
-    private ResponseHandler<String, Object> responseHandler = new BuildImageResponseHandler()
-
     /**
      * Input directory containing the build context. Defaults to "$projectDir/docker".
      */
@@ -118,10 +114,7 @@ class DockerBuildImage extends AbstractDockerRemoteApiTask implements RegistryCr
         }
 
         def response = buildImageCmd.exec(threadContextClassLoader.createBuildImageResultCallback())
-        imageId = responseHandler.handle(response)
-    }
-
-    void setResponseHandler(ResponseHandler<String, Object> responseHandler) {
-        this.responseHandler = responseHandler
+        imageId = response.awaitImageId()
+        logger.quiet "Created image with ID '$imageId'."
     }
 }